add __init__, reorganize import from __init__
