+++ backend/core/repository/brain/set_as_default_brain_for_user.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 
-from models.settings import get_supabase_client
-from repository.brain.get_default_user_brain import get_user_default_brain
+from models import get_supabase_client
+from repository.brain import get_user_default_brain
 
 
 def set_as_default_brain_for_user(user_id: UUID, brain_id: UUID):
