+++ backend/core/repository/chat/get_chat_by_id.py
@@ -1,5 +1,4 @@
-from models.chat import Chat
-from models.settings import get_supabase_db
+from models import Chat, get_supabase_db
 
 
 def get_chat_by_id(chat_id: str) -> Chat:
