+++ backend/core/repository/prompt/update_prompt_by_id.py
@@ -1,8 +1,7 @@
 from uuid import UUID
 
 from models.databases.supabase.prompts import PromptUpdatableProperties
-from models.prompt import Prompt
-from models.settings import get_supabase_db
+from models import Prompt, get_supabase_db
 
 
 def update_prompt_by_id(prompt_id: UUID, prompt: PromptUpdatableProperties) -> Prompt:
