+++ backend/core/repository/user/get_user_id_by_user_email.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from models.settings import get_supabase_client
+from models import get_supabase_client
 
 
 def get_user_id_by_user_email(email: str) -> UUID:
