+++ backend/core/repository/user_identity/__init__.py
@@ -1,0 +1,3 @@
+from .create_user_identity import create_user_identity
+from .get_user_identity import get_user_identity
+from .update_user_identity import update_user_identity, UserIdentityUpdatableProperties
