+++ backend/core/repository/user_identity/create_user_identity.py
@@ -1,5 +1,4 @@
-from models.settings import get_supabase_client
-from models.user_identity import UserIdentity
+from models import get_supabase_client, UserIdentity
 
 
 def create_user_identity(user_identity: UserIdentity) -> UserIdentity:
