+++ backend/core/repository/user_identity/get_user_identity.py
@@ -1,7 +1,6 @@
 from uuid import UUID
 
-from models.settings import get_supabase_client
-from models.user_identity import UserIdentity
+from models import get_supabase_client, UserIdentity
 from repository.user_identity.create_user_identity import create_user_identity
